[ICP-5348] Fix for parameters 80 and 24.

+++ devicetypes/fibargroup/fibaro-motion-sensor-zw5.src/fibaro-motion-sensor-zw5.groovy
@@ -97,7 +97,7 @@ metadata {
 				type: "paragraph",
 				element: "paragraph"
 			)
-			def defVal = it.def
+			def defVal = it.def as Integer
 			def descrDefVal = it.options ? it.options.get(defVal) : defVal
 			input(
 				name: it.key,
